+++ src/main/java/it/live/iqgame/repository/AttemptsRepository.java
@@ -7,5 +7,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.UUID;
 
 public interface AttemptsRepository extends JpaRepository<Attempts, UUID> {
-    Integer countAllByUserIdAndSubjectIdAndQuestionTypeAndIsFinished(Long user_id, Long subject_id, QuestionType questionType, Boolean isFinished);
+    Long countAllByUserIdAndSubjectIdAndQuestionTypeAndIsFinished(Long user_id, Long subject_id, QuestionType questionType, Boolean isFinished);
 }
